refactor(web): rename multisig-account-addresses to match its non-hook export
